Crate quaternion_traits

This library provides traits and many function for dealing with quaternions.

Reason

This library exists to provide two things:

A way to tie in the multiple existing implementations of quaternions in rust.

Add functions for pure maths purpaces in rust (like sqrt, cos and pow)

Curently this library is implemented for:

core

std (feature std)

quaternion

num-complex (feature num-complex)

num-traits (feature num-traits)

Details

Currently the quaternion traits are implemented only in core rust and num, but it’s planned to add (though optional dependencies) these traits to crates like bevy, ggez and the other quaternion crates out there.

Due to how the traits are implemented this crate is naturaly usable with the quaternion crate. So for any crates that use this crate this dependency is (hopefully) frictionless.

If possible this crate should implement functions for every quaternion use. And it should implement every function that a crate it’s comapatble with has.

This crate also currently has these pure maths functions for quaternions (excluding common ones): exp, ln, sqrt, sin, sinh, sec, cos, cosh, csc, sin_cos, tan, tanh, cot, coth.

This crate provides currently an unstable form of these functions:

pow (the equasion used seams to not be fully agreed on though so it’s at risk of change if another equasion comes out that is guaranteed to be correct)

log (simple ln division only works for complex quaternions that include the real axis and only one of the other 3 imaginary axis. Will remain unstable unill a better algorithm is found or it’s found that quaternion logs can not have a formula for whatever reason)

List of features:

unstable: Enables items that may change functionality or may be removed entirely.

std: Adds alloc feature, adds Std struct.

alloc: Adds Quaternion, Vector, Complex and Scalar implementations for Box, Arc, Rc and Cow.

num-traits: (If unstable is enabled) adds Pow to Quat, (if std is enabled) adds Float and all the required traits to the Std struct.

num-complex: Adds num-traits feature, adds Complex implementation for the Complex struct in this crate.

Structs

Quat

The struct representation of the Quaternion trait.

Std

Wrapper that changes the Axis methods from the libm ones to the std ones.

Traits

Axis

A representation of the real number line.

Complex

The general representation for any complex number type.

ComplexConstructor

A constructor for complex numbers.

ComplexConsts

Adds constants associated with any complex number.

Quaternion

The general representation of any quaternion type.

QuaternionConstructor

A constructor for quaternions.

QuaternionConsts

Adds constants associated with any quaternion.

QuaternionMethods

Adds most functions in this crate as methods that take self and/or return Self.

Rotation

The general representation of any rotation baised on euler angles.

RotationConstructor

A constructor for scalar values.

Scalar

The general representation for any scalar type.

ScalarConstructor

A constructor for scalar values.

ScalarConsts

Adds constants associated with any scalar value.

Vector

The general representation for any vector type.

VectorConstructor

A constructor for vectors.

VectorConsts

Adds constants associated with any vectors.

Functions

abs

Gets the absolute value of a quaternion. (Also knows as it’s “length”)

abs_squared

Gets the squared absolute value of a quaternion. (Also knows as it’s squared “length”)

add

Adds two quaternions.

axis_angle

Constructs a quaternion from a given axis unit vector and a given angle.

axis_angle_unchecked⚠

Constructs a quaternion from a given axis unit vector and a given angle.

complex_part

Gets the complex part of a quaternion.

conj

Gets the conjugate of this quaternion.

cos

Calculates the cosinus of a quaternion.

cosh

Calculates the hyperbolic cosinus of a quaternion.

cot

Calculates the cotangent of a quaternion

coth

Calculates the sinus of a quaternion.

csc

Calculates the cosecant of a quaternion.

display

Writes a quaternion representation to a formatter/string.

dist

Gets the distance inbetween the coordenates of two quaternions.

div

Divides a quaternion by another one.

div_reversed

Divides a quaternion by another one in reversed order.

dot

Gets the dot product of two quaternions.

eq

Checks if two types represent the same quaternion.

exp

Raises the number e to a quaternion power.

from_complex

Constructs a quaternion from a complex number representation.

from_rotation

Constructs a quaternion from a rotation.

from_scalar

Constructs a quaternion from a scalar value.

from_vector

Constructs a quaternion from a vector representation.

identity

Constructs the positive real unit quaternion. (Multiplicative identity)

inv

Gets the inverse quaternion of a quaternion.

is_complex

Checks to see if this quaternion is a complex value.

is_near

Checks if the distance between two quaternions is less then Num::ERROR.

is_on_axis_plane

Checks to see if this quaternion is on a axis plane.

is_scalar

Checks to see if this quaternion is only a scalar value.

is_vector

Checks to see if this quaternion is a vector value. (or aure quaternion)

ln

Gets the natural logarithm of a quaternion.

log

Gets the logarithm of a quaternion with a quaternion base.

mul

Multiplies a quaternion to another one.

mul_reversed

Multiplies two quaternions in reversed order.

nan

Constructs a quaternion that has all axies set to Num::NANs.

neg

Gets the negative of this quaternion.

norm

Gets the normal of a quaternion.

origin

Constructs the origin quaternion. (Aditive identity)

pow_f

Raises a quaternion to a scalar power.

pow_i

Raises a quaternion to an integer power.

pow_q

Raises a quaternion to a quaternion power.

pow_u

Raises a quaternion to a positive integer power.

product

Multiplies all the quaternions in an iterator.

rotate_vector

Gives the vector rotated by the given quaternion

rotation_from_to

Constructs a quaternion representing the rotation inbetween two vectors.

scalar_part

Gets the scalar part of a quaternion.

scale

Scales a quaternion.

sec

Calculates the secant of a quaternion.

sin

Calculates the sinus of a quaternion.

sin_cos

Calculates the sinus and cosinus of a quaternion at once.

sinh

Calculates the hyperbolic sinus of a quaternion.

small_abs

Gets the absolute value of a quaternion close to the origin.

sqrt

Calculates the sqaure root of a quaternion.

sub

Subtracts a quaternion from another one.

sum

Adds all the quaternions in an iterator.

tan

Calculates the tangent of a quaternion

tanh

Calculates the sinus of a quaternion.

to_complex

Constructs a complex number representation from a quaternion.

to_rotation

Constructs a rotation representation from a quaternion.

to_scalar

Constructs a scalar value from a quaternion.

to_string

Turns a quaternion representation into a String.

to_vector

Constructs a complex number representation from a quaternion.

unscale

Scales a quaternion by the inverse of the scalar.

vector_part

Gets the vector part of a quaternion.