quaternion_traits

Function small_abs

pub fn small_abs<Num, Out>(quaternion: impl Quaternion<Num>) -> Out
where
    Num: Axis,
    Out: ScalarConstructor<Num>,

Gets the absolute value of a quaternion close to the origin.

Calculates the sqared absolute value of the quaternion multiplied by Num::ERROR to the -2 power. Then it takes the square root. The it multiplies by Num::ERROR to the first power.

Note: The operations above are the rough order in which they are done.