quaternion_traits

Struct Quat

pub struct Quat<Num: Axis = f32, T = (Num, [Num; 3])> {
    pub quat: T,
    /* private fields */
}

The struct representation of the Quaternion trait.

Try not to use this quaternion struct if:

If you don’t mind using the a tuple or an array.

You don’t plan on using any operators like + or * and just functions/traits.

You already have a quaternion type that implements the Quaternion, QuaternionConstructor, QuaternionConsts and QuaternionMethods traits.

Reasoning: This struct exists just as a ease of use if you need a quaternion struct and do not want to make your own or get one from another crate.

Fields
quat: T

The quaternion held by this struct.

Implementations

impl<Num: Axis, T> Quat<Num, T>

pub const fn new(quat: T) -> Self

Creates a new instince of this struct.