quaternion_traits

Struct Std

pub struct Std<Num: Axis>(pub Num);

Wrapper that changes the Axis methods from the libm ones to the std ones.

Tuple Fields
0: Num
Implementations

impl<Num: Axis> Std<Num>

pub const fn new(number: Num) -> Self

Creates a new Std<Num>

pub const fn get(self) -> Num

Gets the nubmer value.

pub fn map(&mut self, change: impl FnOnce(Num) -> Num)

Maps the number value of self into another one.