quaternion_traits

Trait Axis

pub trait Axis:
    Sized
    + Add<Output = Self>
    + Sub<Output = Self>
    + Mul<Output = Self>
    + Div<Output = Self>
    + Neg<Output = Self>
    + PartialOrd
    + PartialEq
    + Copy
    + 'static {
    const ONE: Self;
    const ZERO: Self;
    const TAU: Self;
    const NAN: Self;
    const ERROR: Self;

    // Required methods
    fn is_nan(&self) -> bool;
    fn sqrt(self) -> Self;
    fn pow(self, exp: Self) -> Self;
    fn sin_cos(self) -> (Self, Self);
    fn asin(self) -> Self;
    fn sinh(self) -> Self;
    fn acos(self) -> Self;
    fn cosh(self) -> Self;
    fn atan2(self, bottom: Self) -> Self;
    fn exp(self) -> Self;
    fn ln(self) -> Self;

    // Provided methods
    fn sin(self) -> Self { ... }
    fn cos(self) -> Self { ... }
    fn abs(self) -> Self { ... }

}

A representation of the real number line.

If the type can aproximate real numbers (eg: floating point) then it qualifyes. If it’s limited to a surtun number type (eg: integers) then it does not. For a type to qualify it must represent one dimension line that is as contineous as possible. It does not need to

This is manualy implemented for f32 and f64 by default.

Implementation

If you want to implement this trait for a custom type T make sure the following holds:

T::ZERO == -T::ZERO

T::ZERO < T::ERROR < T::ONE

For any a: T (optionaly for T::NAN) -> a == a

For any a: T -> a + T::ZERO == a

For any a: T -> a * T::ONE == a

For any a: T -> a * T::ZERO == T::ZERO

For any a: T -> T::ZERO - a == -a

For any a: T, b: T -> a + b == b + a

For any a: T, b: T -> a * b == b * a

For any a: T, b: T -> a / b == a * (T::ONE / b)

For any a: T, b: T -> (a + b) - b == a

Required Associated Constants

const ONE: Self

The multiplicative identity.

const ZERO: Self

The additive identity.

const TAU: Self

An aproximation to the circumfrince of a circle divided by it’s radius.

TAU = 2 * PI

const NAN: Self

The representation of a Not a Number value.

const ERROR: Self

Used as the aporximative precision error for flaoting point arithmatic.

Required Methods

fn is_nan(&self) -> bool

Checks to see if self is NAN. (x == Self::NAN is not guaranteed to work)

fn sqrt(self) -> Self

Calculates the square root of self.

fn pow(self, exp: Self) -> Self

Calculates the self raised to the exp power.

fn sin_cos(self) -> (Self, Self)

Calculates the sine and cosine of self at once.

fn asin(self) -> Self

Calculates the arcsine of self.

fn sinh(self) -> Self

Calculates the arcsine of self.

fn acos(self) -> Self

Calculates the arccosine of self.

fn cosh(self) -> Self

Calculates the arccosine of self.

fn atan2(self, bottom: Self) -> Self

Calculates the arctangent of self / bottom.

fn exp(self) -> Self

Calculates e raised to the power of self.

e ≈ 2.71828...

fn ln(self) -> Self

Calculates natural logarithm self.

Provided Methods

fn sin(self) -> Self

Calculates the sine of self.

fn cos(self) -> Self

Calculates the cosine of self.

fn abs(self) -> Self

Calculates the absolute value of self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.