quaternion_traits

Trait Matrix

pub trait Matrix<Num, const N: usize> {
    // Required method
    fn get_unchecked(&self, row: usize, col: usize) -> Num;

    // Provided methods
    fn get(&self, row: usize, col: usize) -> Option<Num> { ... }
    fn to_array(&self) -> [[Num; N]; N] { ... }
}

The general representation of any NxN rotation matrix.

Only Matrix<_, 2>, Matrix<_, 3> and Matrix<_, 4> have impls and are used.

Note: The get_unchecked method is used as if it’s a cheap operation.

Required Methods

fn get_unchecked(&self, row: usize, col: usize) -> Num

Gets the value represented at (row, col)

Important

This value should not panic for values of row and col that are both smaller then N.

Provided Methods

fn get(&self, row: usize, col: usize) -> Option<Num>

Checks if row and col are out of bounds before getting the value at (row, col).

Important

By default this returns None only if row and col are both smaller then N.

fn to_array(&self) -> [[Num; N]; N]

Turns this matrix reprezentation into a NxN array.