quaternion_traits

Trait QuaternionMethods

pub trait QuaternionMethods<Num: Axis>:
    Quaternion<Num>
    + NewQuaternion<Num>
    + Sized {

    // Provided methods
    fn origin() -> Self { ... }
    fn identity() -> Self { ... }
    fn nan() -> Self { ... }
    fn add(&self, other: &impl Quaternion<Num>) -> Self { ... }
    fn sub(&self, other: &impl Quaternion<Num>) -> Self { ... }
    fn mul(&self, other: &impl Quaternion<Num>) -> Self { ... }
    fn mul_reversed(&self, other: &impl Quaternion<Num>) -> Self { ... }
    fn div(&self, other: &impl Quaternion<Num>) -> Self { ... }
    fn scale(&self, other: impl Scalar<Num>) -> Self { ... }
    fn unscale(&self, other: impl Scalar<Num>) -> Self { ... }
    fn is_scalar(&self) -> bool { ... }
    fn is_complex(&self) -> bool { ... }
    fn is_vector(&self) -> bool { ... }
    fn is_on_axis_plane(&self) -> bool { ... }
    fn eq(&self, other: &impl Quaternion<Num>) -> bool { ... }
    fn abs(&self) -> Num { ... }
    fn abs_squared(&self) -> Num { ... }
    fn dot(&self, other: &impl Quaternion<Num>) -> Num { ... }
    fn norm(&self) -> Self { ... }
    fn conj(&self) -> Self { ... }
    fn neg(&self) -> Self { ... }
    fn inv(&self) -> Self { ... }
    fn is_near(&self, other: &impl Quaternion<Num>) -> bool { ... }
    fn dist(&self, other: &impl Quaternion<Num>) -> Num { ... }
    fn sqrt(&self) -> Self { ... }
    fn pow_i(&self, exp: i32) -> Self { ... }
    fn pow_u(&self, exp: u32) -> Self { ... }
    fn pow_f(&self, exp: &impl Scalar<Num>) -> Self { ... }
    fn exp(&self) -> Self { ... }
    fn ln(&self) -> Self { ... }
    fn sin(&self) -> Self { ... }
    fn sinh(&self) -> Self { ... }
    fn sec(&self) -> Self { ... }
    fn cos(&self) -> Self { ... }
    fn cosh(&self) -> Self { ... }
    fn csc(&self) -> Self { ... }
    fn tan(&self) -> Self { ... }
    fn tanh(&self) -> Self { ... }
    fn vector_part(&self) -> Self { ... }
    fn complex_part(&self) -> Self { ... }
    fn scalar_part(&self) -> Self { ... }
    fn to_vector<V: NewVector<Num>>(&self) -> V { ... }
    fn to_complex<C: NewComplex<Num>>(&self) -> C { ... }
    fn to_scalar<S: NewScalar<Num>>(&self) -> S { ... }
    fn to_rotation<R: NewRotation<Num>>(&self) -> R { ... }
    fn from_vector<V: Vector<Num>>(vector: &V) -> Self { ... }
    fn from_complex<C: Complex<Num>>(complex: &C) -> Self { ... }
    fn from_scalar<S: Scalar<Num>>(scalar: &S) -> Self { ... }
    fn from_rotation<R: Rotation<Num>>(rotation: &R) -> Self { ... }

}

Adds most functions in this crate as methods that take self and/or return Self.

Most methods have dry documentation and just link to their respective crate function. If possible it’s planned to have the same* documentation for both instances of each function in the future.

* With examples that match their respective function.

Provided Methods

fn origin() -> Self

Constructs the origin quaternion. (Additive identity)

Example

use quaternion_traits::QuaternionMethods;
 
let quat: [f32; 4] = QuaternionMethods::<f32>::origin();
 
assert_eq!( quat, [0.0, 0.0, 0.0, 0.0] );

fn identity() -> Self

Constructs the positive real unit quaternion. (Multiplicative identity)

Example

use quaternion_traits::QuaternionMethods;
 
let quat: [f32; 4] = QuaternionMethods::<f32>::identity();
 
assert_eq!( quat, [1.0, 0.0, 0.0, 0.0] );

fn nan() -> Self

Constructs a quaternion that has all axies set to Num::NANs.

Example

use quaternion_traits::QuaternionMethods;
 
let quat: [f32; 4] = QuaternionMethods::<f32>::nan();
 
assert!( quat[0].is_nan() );
assert!( quat[1].is_nan() );
assert!( quat[2].is_nan() );
assert!( quat[3].is_nan() );

fn add(&self, other: &impl Quaternion<Num>) -> Self

Adds two quaternions togheder.

Check the add function in the root for more info.

fn sub(&self, other: &impl Quaternion<Num>) -> Self

Subtracts a quaternion from another one.

Check the sub function in the root for more info.

fn mul(&self, other: &impl Quaternion<Num>) -> Self

Muliplies a quaternion to another one.

Check the mul function in the root for more info.

fn mul_reversed(&self, other: &impl Quaternion<Num>) -> Self

Muliplies a quaternion to another one in a reversed order.

Check the mul_reversed function in the root for more info.

fn div(&self, other: &impl Quaternion<Num>) -> Self

Divides a quaternion from another one.

Check the div function in the root for more info.

fn scale(&self, other: impl Scalar<Num>) -> Self

Scales a quaternion.

Equivalent to multiplying a quaternion by a scalar quaternion.

Check the scale function in the root for more info.

fn unscale(&self, other: impl Scalar<Num>) -> Self

Scales a quaternion by the inverse of the scalar.

Equivalent to dividing a quaternion by a scalar quaternion.

Check the unscale function in the root for more info.

fn is_scalar(&self) -> bool

Checks if a quaternion is a scalar value.

Check the is_scalar function in the root for more info.

fn is_complex(&self) -> bool

Checks if a quaternion is a complex value.

Check the is_complex function in the root for more info.

fn is_vector(&self) -> bool

Checks if a quaternion is a vector value.

Check the is_vector function in the root for more info.

fn is_on_axis_plane(&self) -> bool

Checks if a quaternion is on an axis plane.

Check the is_on_axis_plane function in the root for more info.

fn eq(&self, other: &impl Quaternion<Num>) -> bool

Checks if two quaternion represent the same value.

Check the eq function in the root for more info.

fn abs(&self) -> Num

Gets the absolute value of a quaternion. (Also knows as it’s “length”)

Check the abs function in the root for more info.

fn abs_squared(&self) -> Num

Gets the squared absolute value of a quaternion. (Also knows as it’s squared “length”)

Check the abs_squared function in the root for more info.

fn dot(&self, other: &impl Quaternion<Num>) -> Num

Gets the dot product of two quaternions.

Check the dot function in the root for more info.

fn norm(&self) -> Self

Gets the normal of a quaternion.

Check the norm function in the root for more info.

fn conj(&self) -> Self

Gets the conjugate of a quaternion.

Check the conj function in the root for more info.

fn neg(&self) -> Self

Gets the negative of a quaternion.

Check the neg function in the root for more info.

fn inv(&self) -> Self

Gets the inverse of a quaternion.

Check the inv function in the root for more info.

fn is_near(&self, other: &impl Quaternion<Num>) -> bool

Checks if the distance inbetween two quaternions is less then Num::ERROR.

Check the is_near function in the root for more info.

fn dist(&self, other: &impl Quaternion<Num>) -> Num

Gets the distance inbetween the coordonates of two quaternions.

Check the dist function in the root for more info.

fn sqrt(&self) -> Self

Gets the square root of a quaternion.

Check the sqrt function in the root for more info.

fn pow_i(&self, exp: i32) -> Self

Raises a quaternion to an integer power.

Check the pow_i function in the root for more info.

fn pow_u(&self, exp: u32) -> Self

Raises a quaternion to a positive integer power.

Check the pow_u function in the root for more info.

fn pow_f(&self, exp: &impl Scalar<Num>) -> Self

Raises a quaternion to a scalar power.

Check the pow_f function in the root for more info.

fn exp(&self) -> Self

Raises the number e to a quaternion power.

Check the exp function in the root for more info.

fn ln(&self) -> Self

Gets the natural logarithm of a quaternion.

Check the ln function in the root for more info.

fn sin(&self) -> Self

Gets the sinus of a quaternion.

Check the sin function in the root for more info.

fn sinh(&self) -> Self

Gets the hyperbolic sinus of a quaternion.

fn sec(&self) -> Self

Gets the secant of a quaternion.

fn cos(&self) -> Self

Gets the cosinus of a quaternion.

Check the cos function in the root for more info.

fn cosh(&self) -> Self

Gets the hyperbolic cosinus of a quaternion.

fn csc(&self) -> Self

Gets the cosecant of a quaternion.

fn tan(&self) -> Self

Gets the tangent of a quaternion.

Check the tan function in the root for more info.

fn tanh(&self) -> Self

Gets the hyperbolic tangent of a quaternion.

fn vector_part(&self) -> Self

Gets the vector part of a quaternion.

Check the vector_part function in the root for more info.

fn complex_part(&self) -> Self

Gets the complex part of a quaternion.

Check the complex_part function in the root for more info.

fn scalar_part(&self) -> Self

Gets the scalar part of a quaternion.

Check the scalar_part function in the root for more info.

fn to_vector<V: NewVector<Num>>(&self) -> V

Turns a quaternion representation into a vector value representation.

Check the to_vector function in the root for more info.

fn to_complex<C: NewComplex<Num>>(&self) -> C

Turns a quaternion representation into a complex number representation.

Check the to_complex function in the root for more info.

fn to_scalar<S: NewScalar<Num>>(&self) -> S

Turns a quaternion representation into a scalar representation.

Check the to_scalar function in the root for more info.

fn to_rotation<R: NewRotation<Num>>(&self) -> R

Turns a quaternion representation into a rotation.

Check the to_rotation function in the root for more info.

fn from_vector<V: Vector<Num>>(vector: &V) -> Self

Constructs a quaternion representation from a vector.

Check the from_vector function in the root for more info.

fn from_complex<C: Complex<Num>>(complex: &C) -> Self

Constructs a quaternion representation from a complex number.

Check the from_complex function in the root for more info.

fn from_scalar<S: Scalar<Num>>(scalar: &S) -> Self

Constructs a quaternion representation from a scalar value.

Check the from_scalar function in the root for more info.

fn from_rotation<R: Rotation<Num>>(rotation: &R) -> Self

Constructs a unit quaternion representation from a rotation.

Check the from_rotation function in the root for more info.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.