quaternion_traits

Trait Rotation

pub trait Rotation<Num: Axis> {
    // Required methods
    fn roll(&self) -> Num;
    fn pitch(&self) -> Num;
    fn yaw(&self) -> Num;
}

The general representation of any rotation baised on euler angles.

This struct uses radians (for all prodived implementations of Axis). No actual mesurment is inforced by the struct it’self, so if you wish to use degrees you can, you just eather need a wrapper or to modify the values each time you use it.

Note: The roll, pitch and yaw methods are used as if they are cheap operations.

Required Methods

fn roll(&self) -> Num

The roll of this rotation. (angle on the z axis)

fn pitch(&self) -> Num

The pitch of this rotation. (angle on the y axis)

fn yaw(&self) -> Num

The yaw of this rotation. (angle on the x axis)