quaternion_traits

Trait Scalar

pub trait Scalar<Num: Axis> {
    // Required method
    fn scalar(&self) -> Num;
}

The general representation for any scalar type.

Marks that this type can turn into an Axis type.

Note: The scalar method is used as if its a cheap operation.

Required Methods

fn scalar(&self) -> Num

The Axis representation of this scalar value.